backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
python-dotenv==1.0.1
httpx==0.28.1
requests>=2.28.0
pydantic==2.10.3
pydantic-settings==2.6.1
numpy==2.0.2
scikit-learn==1.6.0
joblib==1.4.2
pymongo==4.10.1
google-genai==0.8.3
google-generativeai>=0.8.0
# Python 3.9 backport for X | Y union type hints with Pydantic v2
eval_type_backport>=0.2.0
# File upload support (multipart/form-data)
python-multipart>=0.0.9
# PDF text extraction
pypdf>=4.0.0
# ML / embeddings
scikit-learn>=1.4.0
# Settings management
pydantic-settings>=2.0.0
# Async HTTP client
httpx>=0.27.0
# Async MongoDB driver
motor>=3.3.0
# NVIDIA NV-Embed-v2 via SentenceTransformer (local fallback)
# Requires ~3 GB model download on first use:
sentence-transformers>=3.0.0
uvicorn
fastapi
einops
motor
datasets    
google-genai
openai